% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_challengers.R
\name{apply_pca_transform}
\alias{apply_pca_transform}
\title{Apply a training PCA transform}
\usage{
apply_pca_transform(fit, data)
}
\arguments{
\item{fit}{PCA preprocessing object.}

\item{data}{New data.}
}
\value{
Data frame of retained principal components.
}
\description{
Apply a training PCA transform
}
