% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_challengers.R
\name{apply_quantile_bins}
\alias{apply_quantile_bins}
\title{Apply training quantile bins}
\usage{
apply_quantile_bins(fit, data)
}
\arguments{
\item{fit}{Quantile-bin preprocessing object.}

\item{data}{New data.}
}
\value{
Data frame of fixed-level factors.
}
\description{
Apply training quantile bins
}
