% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{canonical_target_leagues}
\alias{canonical_target_leagues}
\title{Return the canonical target leagues}
\usage{
canonical_target_leagues(config_path = .league_config_path())
}
\arguments{
\item{config_path}{Path to the league configuration.}
}
\value{
Character vector of seven target leagues.
}
\description{
Return the canonical target leagues
}
