% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_evaluation.R
\name{discrete_crps}
\alias{discrete_crps}
\title{Calculate discrete CRPS for a probability mass function}
\usage{
discrete_crps(pmf, observed)
}
\arguments{
\item{pmf}{Probabilities for integer outcomes starting at zero.}

\item{observed}{Observed non-negative integer outcome.}
}
\value{
Scalar CRPS.
}
\description{
Calculate discrete CRPS for a probability mass function
}
