% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_dynamic_ratings.R
\name{dynamic_team_model_features}
\alias{dynamic_team_model_features}
\title{Return the dynamic team model feature set}
\usage{
dynamic_team_model_features()
}
\value{
Character vector used by regularized challenger models.
}
\description{
Return the dynamic team model feature set
}
