% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_challengers.R
\name{fit_quantile_bins}
\alias{fit_quantile_bins}
\title{Learn quantile bins from training data only}
\usage{
fit_quantile_bins(data, feature_names, bins = 5L)
}
\arguments{
\item{data}{Training data.}

\item{feature_names}{Numeric features.}

\item{bins}{Requested number of bins.}
}
\value{
Quantile-bin preprocessing object.
}
\description{
Learn quantile bins from training data only
}
