% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_schema.R
\name{required_oe_columns}
\alias{required_oe_columns}
\title{Required Oracle's Elixir columns}
\usage{
required_oe_columns()
}
\value{
Character vector of canonical source column names.
}
\description{
Required Oracle's Elixir columns
}
