% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_model.R
\name{summarize_mcmc_diagnostics}
\alias{summarize_mcmc_diagnostics}
\title{Summarize mandatory MCMC diagnostics}
\usage{
summarize_mcmc_diagnostics(fit)
}
\arguments{
\item{fit}{CmdStanMCMC object.}
}
\value{
One-row diagnostic summary.
}
\description{
Summarize mandatory MCMC diagnostics
}
