% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manifest.R
\name{validate_data_manifest}
\alias{validate_data_manifest}
\title{Validate an Oracle's Elixir data manifest}
\usage{
validate_data_manifest(manifest, raw_dir)
}
\arguments{
\item{manifest}{Manifest data frame.}

\item{raw_dir}{Directory containing the files.}
}
\value{
\code{TRUE} invisibly when valid.
}
\description{
Validate an Oracle's Elixir data manifest
}
