% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oe_schema.R
\name{validate_oe_schema}
\alias{validate_oe_schema}
\title{Validate Oracle's Elixir source columns}
\usage{
validate_oe_schema(columns)
}
\arguments{
\item{columns}{Character vector of column names.}
}
\value{
\code{TRUE} invisibly when valid.
}
\description{
Validate Oracle's Elixir source columns
}
