% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manifest.R
\name{write_data_manifest}
\alias{write_data_manifest}
\title{Write a data manifest as YAML}
\usage{
write_data_manifest(manifest, path)
}
\arguments{
\item{manifest}{Valid manifest data frame.}

\item{path}{Destination YAML path.}
}
\value{
Destination path invisibly.
}
\description{
Write a data manifest as YAML
}
